fix: always use `url` field for image
Previously the `hdurl` field was used, but Discord would not always display the
image. Using the non-high-res version of the image is fine, since it's not gonna
take up a lot of space on the screen.

An alternative is to check if the image is too big, and only use `url` in that
case. This is a more complicated solution that depends on undocumented Discord
behavior, such as the point where it refuses to show the image, for very little
extra benefit.